+++ examples/gui_direct2.py
@@ -92,7 +92,11 @@ class Loop:
                     print(err)
 
             # draw a frame
-            draw_frame()
+            try:
+                draw_frame()
+            except wgpu.DrawCancelled:  # window occluded
+                time.sleep(0.1)
+                continue
             # present the frame to the screen
             context.present()
             # stats