+++ wgpu/_classes.py
@@ -23,6 +23,7 @@ from . import flags, enums, structs
 
 __all__ = [
     "GPU",
+    "DrawCancelled",
     "GPUAdapter",
     "GPUAdapterInfo",
     "GPUBindGroup",
@@ -2447,6 +2448,13 @@ class GPUDeviceLostInfo:
     def reason(self) -> enums.DeviceLostReasonEnum:
         """The reason (enums.GPUDeviceLostReason) for the device getting lost. Can be None."""
         return self._reason
+
+
+@apidiff.add("Added because Python deals with native windows")
+class DrawCancelled(Exception):  # noqa: N818
+    """An excepion raised to cancel a draw when the surface texture could not be obtained,
+    e.g. because the window is hidden or occluded.
+    """
 
 
 class GPUError(Exception):